Module spinach::lattice

All lattice merges.

Structs

DominatingPair
A composed lattice. Given two lattice merges AF and BF, merge in lexicographical order.

Hide

Intersect
Set intersection lattice.

MapUnion
Map union composed lattice.

Max
For totally-ordered demains, take the max value.

Min
For totally-ordered demains, take the min value.

RangeToZeroI32
Mingwei’s weird semilattice. Lattice is defined as, given signed integers A and B, take the value in the range [A, B] (or [B, A]) which is closest to zero. (Note that in general this will be A, B, or zero).

RefOptional
Wrap an existing lattice in Option, where None is smaller than all other elements.

Union
Classic set union lattice.

Traits

Lattice
Lattice trait, a merge function which defines a lattice.