Trait spinach::lattice::Lattice

pub trait Lattice {
    type Domain: Clone;
    fn merge_in(val: &mut Self::Domain, delta: Self::Domain);

    fn partial_cmp(val: &Self::Domain, delta: &Self::Domain) -> Option<Ordering>;

    fn delta(val: &Self::Domain, delta: &mut Self::Domain) -> bool;
}

Lattice trait, a merge function which defines a lattice.

Associated Types

type Domain: Clone

Loading content...
Required methods

fn merge_in(val: &mut Self::Domain, delta: Self::Domain)

fn partial_cmp(val: &Self::Domain, delta: &Self::Domain) -> Option<Ordering>

fn delta(val: &Self::Domain, delta: &mut Self::Domain) -> bool

Given an current VAL, modify delta so it is of a minimum size. Return TRUE if delta is non-trivial. Return FALSE if delta can be ignored (VAL dominates DELTA).

Loading content...