Function spinach::op::handoff_op

pub fn handoff_op<F: Flow, T>() -> (HandoffPushOp<F, T>, HandoffPullOp<F, T>)

Create a connected sending and receiving handoff pair, HandoffPushOp and HandoffPullOp.

This is meant as a lightweight single-threaded single-item-buffered alternative to channel_op.

Supports both Df and Rx flows.

TODO: NOTE: Rx flows end up in the single element buffer, so they may unneccearily block.